[qren] moved constants to Const.pm, use main()
Cleaned up the code a bit by moving constants to a Const.pm module so
they are more centrally located. In the future, if we want to move
these settings into a JSON file, we can easily update Const.pm to do
the loading.

Also updated code to use a main() function.