[qren] auto process iphone tags, show config option
Updated code to automatically process iphone tags to handle the "Plus"
part in name, so future iphone models should be handled automatically.

Can't store the substitution rule in config file because it requires
the use of eval which is dangerious, so put it in code.

Also added option -g to print config file content, useful for when we
add new config file entries to see how the program reads it.