[qren] support 'w' and 'd' in offset unit specification.
This change allows easier adjustment for gps rollover offset, such as
"-o1024w" for adding 2014 weeks to the timestammp.

Note that the before and after date may have different daylight saving
time value, so pay attention to see if an additional 1h offset is
needed to restore the time to the original value.

Best to rename first using the wrong GPS date, then apply the 1024w
offset, and ensure the hour value didn't change.